requirements.txt
torch
transformers
googletrans==4.0.0-rc1
gtts
Pillow
contextlib2
